+++ custom_components/smartthinq_sensors/switch.py
@@ -21,10 +21,11 @@ from homeassistant.helpers.dispatcher import async_dispatcher_connect
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
-from . import LGEDevice
-from .const import DOMAIN, LGE_DEVICES, LGE_DISCOVERY_NEW
+from .const import LGE_DISCOVERY_NEW
 from .device_helpers import STATE_LOOKUP, LGEBaseDevice
+from .lge_device import LGEDevice
 from .official_control import async_call_official_turn_off, async_call_official_turn_on
+from .runtime_data import get_lge_devices
 from .wideq import (
     WM_DEVICE_TYPES,
     AirConditionerFeatures,
@@ -184,8 +185,7 @@ async def async_setup_entry(
     hass: HomeAssistant, entry: ConfigEntry, async_add_entities: AddEntitiesCallback
 ) -> None:
     """Set up the LGE switch."""
-    entry_config = hass.data[DOMAIN]
-    lge_cfg_devices = entry_config.get(LGE_DEVICES)
+    lge_cfg_devices = get_lge_devices(hass)
 
     _LOGGER.debug("Starting LGE ThinQ switch setup")
